Fix/code status

+++ src/main/java/com/unsis/scunsis_backend/service/activity/ActivityService.java
@@ -3,6 +3,9 @@ package com.unsis.scunsis_backend.service.activity;
 import java.util.List;
 import java.util.Optional;
 
+import com.unsis.scunsis_backend.constants.Constant;
+import com.unsis.scunsis_backend.exception.AppException;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import com.unsis.scunsis_backend.dto.request.activity.ActivityRequest;
@@ -22,6 +25,9 @@ public class ActivityService {
     private final ActivityMapper activityMapper;
 
     public ActivityResponse getById(long activityId){
+        if(activityRepository.existsById(activityId)){
+            throw new AppException(Constant.NOT_FOUND_BY_ID, HttpStatus.NOT_FOUND);
+        }
         Optional<Activity> activity = activityRepository.findById(activityId);
         ActivityResponse activityResponse = activityMapper.toDto(activity.get());
         return activityResponse;
@@ -33,11 +39,16 @@ public class ActivityService {
     }
 
     public void createActivity(ActivityRequest request){
+        if(activityRepository.existsById(request.getActivityId())){
+            throw new AppException(Constant.NOT_FOUND_BY_ID, HttpStatus.NOT_FOUND);
+        }
         activityRepository.save(activityMapper.toEntity(request));
     }
 
     public void deleteById(long activityId){
-        // Necesitamos primero confirmar o buscar que exista esta actividad antes de borrar
+        if(!activityRepository.existsById(activityId)){
+            throw new AppException(Constant.NOT_FOUND_BY_ID, HttpStatus.NOT_FOUND);
+        }
         activityRepository.deleteById(activityId);
     }
