+++ src/main/java/com/unsis/scunsis_backend/service/sender/SenderService.java
@@ -1,14 +1,15 @@
 package com.unsis.scunsis_backend.service.sender;
 
 import java.util.List;
-import java.util.Optional;
 
+import com.unsis.scunsis_backend.constants.Constant;
+import com.unsis.scunsis_backend.exception.AppException;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import com.unsis.scunsis_backend.dto.request.sender.SenderRequest;
 import com.unsis.scunsis_backend.dto.response.sender.SenderResponse;
 import com.unsis.scunsis_backend.mapper.sender.SenderMapper;
-import com.unsis.scunsis_backend.model.sender.Sender;
 import com.unsis.scunsis_backend.repository.sender.ISenderRepository;
 
 import lombok.Data;
@@ -22,24 +23,26 @@ public class SenderService {
     private final SenderMapper senderMapper;
 
     public SenderResponse getById(long senderId){
-        Optional<Sender> sender = senderRepository.findById(senderId);
-        SenderResponse senderResponse = senderMapper.toDto(sender.get());
-        return senderResponse;
+        return senderRepository.findById(senderId)
+                .map(senderMapper::toDto)
+                .orElseThrow(() -> new AppException(Constant.NOT_FOUND_SENDER, HttpStatus.NOT_FOUND));
     }
 
     public List<SenderResponse> getAll(){
-        List<Sender> senders = senderRepository.findAll();
-        List<SenderResponse> sendersResponse = senderMapper.toDtos(senders);
-        return sendersResponse;
+        return senderMapper.toDtos(senderRepository.findAll());
     }
 
     public void deleteById(long senderId){
-        // Validamos si existe(lo buscamos)
+        if(!senderRepository.existsById(senderId)){
+            throw new AppException(Constant.NOT_FOUND_SENDER, HttpStatus.NOT_FOUND);
+        }
         senderRepository.deleteById(senderId);
     }
 
     public void createSender(SenderRequest request){
-        // Validar si existe
+        if(senderRepository.existsByName(request.getName().trim())){
+           throw new AppException(Constant.SENDER_EXISTS, HttpStatus.CONFLICT);
+        }
         senderRepository.save(senderMapper.toEntity(request));
     }
 }
